+++ jdaviz/configs/default/plugins/__init__.py
@@ -12,3 +12,4 @@ from .export.export import *  # noqa
 from .plot_options.plot_options import *  # noqa
 from .markers.markers import * # noqa
 from .data_quality.data_quality import *  # noqa
+from .about.about import *  # noqa